requirements.txt
# Основные зависимости Django
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-gis>=1.0
django-cors-headers>=4.3.0
django-filter>=23.0
django-celery-beat>=2.5.0

# Геопространственные библиотеки
# GDAL>=3.4.0
pyproj>=3.6.0
Shapely>=2.0.0
Fiona>=1.9.0
rasterio>=1.3.0
folium>=0.14.0

# Обработка данных
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Celery и зависимости
celery>=5.3.0
redis>=5.0.0
kombu>=5.3.0

# База данных
psycopg2-binary>=2.9.0

# Утилиты
python-dateutil>=2.8.0
requests>=2.31.0
urllib3>=2.0.0

# Логирование и мониторинг
python-json-logger>=2.0.0

# Форматы данных
geojson>=3.0.0
PyYAML>=6.0.0

# Тестирование
pytest>=7.0.0
pytest-django>=4.5.0
factory-boy>=3.3.0

# Документация
drf-yasg>=1.21.0

# Безопасность
django-environ>=0.11.0
python-dotenv>=1.0.0 